VFB-314 refactoring
